Cast `iv_length` to an int so the comparison can succeed.
The comparison of `iv_length` and `SHA256_DIGEST_LENGTH` will fail even
in a success scenario because `iv_length` is an int and
`SHA256_DIGEST_LENGTH` is an `unsigned int`. This caused the call to the
`run` method on the `OpenSSLSHA256` kernel to fail with a SystemError
since the function simply returns `NULL` instead of creating/raising an
Exception.